+++ sdk/resourcemanager/confluent/armconfluent/zz_generated_validations_client.go
@@ -38,7 +38,7 @@ func NewValidationsClient(subscriptionID string, credential azcore.TokenCredenti
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	ep := cloud.AzurePublic.Services[cloud.ResourceManager].Endpoint
 	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
 		ep = c.Endpoint
 	}
@@ -56,6 +56,7 @@ func NewValidationsClient(subscriptionID string, credential azcore.TokenCredenti
 
 // ValidateOrganization - Organization Validate proxy resource
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2021-12-01
 // resourceGroupName - Resource group name
 // organizationName - Organization resource name
 // body - Organization resource model
@@ -98,7 +99,7 @@ func (client *ValidationsClient) validateOrganizationCreateRequest(ctx context.C
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2021-12-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, body)
 }